Rename LedgerForm to LedgerTable
