+++ client/packages/system/src/Stock/Components/Ledger/index.ts
@@ -1,1 +1,1 @@
-export * from './LedgerForm';
+export * from './LedgerTable';
